Convert scripts from CommonJS to modern ESM format (#7230)
